fixed error with hibernate

+++ android tutorials/Project/FriendFinder 1.0/backendfriendfinder2/src/main/java/friendfinder/persistence/AccountRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by grace on 24/06/17.
  */
-public interface AccountRepository extends CrudRepository<Account, Long> {
+public interface AccountRepository extends CrudRepository<Account, Integer> {
 
     public Account findByEmail(String email);
 }
